[CHS-3]: small style fixes
