+++ src/main/java/mitrofanov82/edu/javagroup1/angry_chess/shared_model/GameResultType.java
@@ -1,5 +1,5 @@
 package mitrofanov82.edu.javagroup1.angry_chess.shared_model;
 
 public enum GameResultType {
-	WHITE_WIN, BLACK_WIN, DRAW, NOT_FINISHED, END_GAME
+	WHITE_WIN, BLACK_WIN, DRAW, IN_PROGRESS, END_GAME_NO_RESULT
 }
