--- src/main/java/mitrofanov82/edu/javagroup1/angry_chess/shared_model/IChessPersistenceAPI.java
@@ -1,8 +1,0 @@
-package mitrofanov82.edu.javagroup1.angry_chess.shared_model;
-
-import java.util.List;
-
-public interface IChessPersistenceAPI {
-	void saveGameResults(IGame game);
-	List<IGame> readGameResults(IPlayer p1, IPlayer p2);
-}
